delete serializer.py and views.py files
create serializers and view folders

update the urls in users app
