+++ users/serializers/users_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from django.contrib.auth.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):  # User GET request serializer
+    class Meta:
+        model = User
+        fields = ['id', 'username', 'email']
